--- src/components/Dashboard/DashboardAdmin/index.ts
@@ -1,1 +1,0 @@
-export { default } from './DashboardAdmin'
