--- src/components/Dashboard/DashboardUser/index.ts
@@ -1,1 +1,0 @@
-export { default } from './DashboardUser'
