--- src/components/Dashboard/index.ts
@@ -1,4 +1,0 @@
-import DashboardAdmin from './DashboardAdmin'
-import DashboardUser from './DashboardUser'
-
-export { DashboardAdmin, DashboardUser }
