+++ src/components/Modals/ModuleModals/ModuleContentsModal/steps/StepThree.tsx
@@ -38,7 +38,9 @@ export const StepThree = ({}: StepProps) => {
         () =>
             users?.results
                 .filter(user => user.UUID !== values.Owner_2_UUID)
-                .filter(user => filterRoles?.includes(user.Rol))
+                .filter(user =>
+                    user.Roles?.some(role => filterRoles?.includes(role))
+                )
                 .map(user => ({
                     label: user.Gebruikersnaam,
                     value: user.UUID,
@@ -50,7 +52,9 @@ export const StepThree = ({}: StepProps) => {
         () =>
             users?.results
                 .filter(user => user.UUID !== values.Owner_1_UUID)
-                .filter(user => filterRoles?.includes(user.Rol))
+                .filter(user =>
+                    user.Roles?.some(role => filterRoles?.includes(role))
+                )
                 .map(user => ({
                     label: user.Gebruikersnaam,
                     value: user.UUID,
